get_role dependency to retrieve user role on a specific dataset. Moved authorization dependencies to standalone module to avoid circular dependencies.
